cleanup: move log functions to util/log.go
There is a util/log.go file, so the log functions in util/util.go can be
moved there.
